move firmware into its own package and share ImmutableBaseModel via common
