+++ Makefile
@@ -5,7 +5,7 @@ all: docs
 proto:
 	cd src && python -m grpc_tools.protoc -I. \
 		--python_out=. --pyi_out=. --grpc_python_out=. \
-		benlink/_benshikj.proto
+		benlink/firmware/_benshikj.proto
 
 docs:
 	pdoc ./src/benlink -o docs --logo /logo.svg
